# Optimize step_two