This PR is not ready for merge, and is meant as a set of guidelines on what we can do to optimize the code even further. We should probably remove the `func_t` being passed as it will probably not be needed, but I've left it there for now.

This PR starts from `step_two` and optimizes it so that it has better cache locality, plus other optimizations.

The current speed-ups I get compared to `step_two` is the following (roughly 3x speedup compared to master for batch size 1, 2x for batch size 32):

<details>

<summary>
Current master (step_two):
</summary>

```
Input tensor: [1, 3, 320, 320]
Num threads: 6

- Check consistency (downsampling to 256x256): OK

- Check consistency (upsampling to 512x512): OK

- Bench upsample_bilinear2d_cpu (5000 rounds) - downsampling to 256x256
Elapsed time: 0.000365611

- Bench ti_upsample_bilinear2d_cpu (5000 rounds) - downsampling to 256x256
Elapsed time: 0.000225859

- Bench upsample_bilinear2d_cpu (5000 rounds) - upsampling to 512x512
Elapsed time: 0.00139943

- Bench ti_upsample_bilinear2d_cpu (5000 rounds) - upsampling to 512x512
Elapsed time: 0.000663689

1 - Benchmark test size as in https://github.com/mingfeima/op_bench-py
Input tensor: [32, 128, 64, 64]
Input is_contiguous memory_format torch.channels_last: 1
Input is_contiguous : 0

- Bench upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0409884

- Bench ti_upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.111906

2 - Benchmark test size as in https://github.com/mingfeima/op_bench-py
Input tensor: [32, 128, 64, 64]
Input is_contiguous memory_format torch.channels_last: 0
Input is_contiguous : 1

- Bench upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0947265

- Bench ti_upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0857989
```

</details>


<details>

<summary>
This PR

</summary>

```
Input tensor: [1, 3, 320, 320]
Num threads: 6

- Check consistency (downsampling to 256x256): OK

- Check consistency (upsampling to 512x512): OK

- Bench upsample_bilinear2d_cpu (5000 rounds) - downsampling to 256x256
Elapsed time: 0.00037262

- Bench ti_upsample_bilinear2d_cpu (5000 rounds) - downsampling to 256x256
Elapsed time: 8.38583e-05

- Bench upsample_bilinear2d_cpu (5000 rounds) - upsampling to 512x512
Elapsed time: 0.00142657

- Bench ti_upsample_bilinear2d_cpu (5000 rounds) - upsampling to 512x512
Elapsed time: 0.00022749

1 - Benchmark test size as in https://github.com/mingfeima/op_bench-py
Input tensor: [32, 128, 64, 64]
Input is_contiguous memory_format torch.channels_last: 1
Input is_contiguous : 0

- Bench upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0423726

- Bench ti_upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0672676

2 - Benchmark test size as in https://github.com/mingfeima/op_bench-py
Input tensor: [32, 128, 64, 64]
Input is_contiguous memory_format torch.channels_last: 0
Input is_contiguous : 1

- Bench upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0927739

- Bench ti_upsample_bilinear2d_cpu (500 rounds) - upsampling to 128x128
Elapsed time: 0.0552326
```

</details>

The commit history is not 100% clean, but should represent different optimizations (or refactorings). Here is a list of what worked:
- special-case the iteration when input stride = 0 and iy stride = 0 -- This should always happen in our current implementation of TensorIterator, but might be good to keep the default case just for safety.
- reorder the computations in the function so that it performs less multiply-adds. So instead of doing `w1 * w2 * x00 + w3 * w2 * x01` I do instead `w2 * (w1 * x00 + w3 * x01)`.
- use a small temporary buffer (of 4 elements) and make the interpolation separable. This gives better cache locality, and provides good speedups
- use `int32_t` for indices. The original reason I did this was to be able to fully exploit vectorization, but seems to give some speed improvements as well. Note that we would need to use `int64_t` indices if the tensors are large, so I would double-check the improvements of using `int32_t` before using it.

Somewhat surprisingly, my first attempt of using vectorized operators yielded slowdowns compared to the past implementation, so I commented it out but left it here for completeness.